refactor: move mapping to controller, update configs, remove CarService interface
- Moved DTO mapping logic to controller level
- Extracted configuration values to application.yml or constants
- Removed CarService interface for simplification
- Changed variable type for 'rate' in Driver entity

+++ driver-service/src/main/java/com/software/modsen/driverservice/model/Driver.java
@@ -21,6 +21,6 @@ public class Driver {
     private String sex;
     private Long car_id;
     private boolean available;
-    private int rate;
+    private double rate;
     private int ratingCount;
 }
